fixed bug, renamed and style code
